move classificador/util to util
