--- classificador/requirements.txt
@@ -1,3 +1,0 @@
-torch
-torchvision
-gdown
