+++ qeep/util/storage.py
